Main Route'/invoice_po/requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.8
pdf2image==1.16.3
google-generativeai==0.4.0
Werkzeug==2.3.7
Pillow==9.3.0
